+++ src/CTLD_crate.lua
@@ -1086,7 +1086,7 @@ function CTLDCrateManager:checkHoverStatus()
     local minH        = ctld.gs("minimumHoverHeight")   or 7.5
     local maxH        = ctld.gs("maximumHoverHeight")   or 12.0
     local hoverTime   = ctld.gs("hoverTime")            or 10
-    local maxSpeed    = ctld.gs("maxSlingloadSpeed")    or 50
+    local maxSpeed    = ctld.gs("maxSlingloadSpeed")    or 26
 
     local players = CTLDPlayerManager.getInstance()._players